refactor: api service
